Continue breaking up Unison.Codebase.Editor.HandleInput into smaller modules
